# Change default of visualize_expected_performance to True