--- a/slothy/core/config.py
+++ b/slothy/core/config.py
@@ -1088,7 +1088,7 @@ class Config(NestedPrint, LockAttributes):
         # Visualization
         self.indentation = 8
         self.visualize_reordering = True
-        self.visualize_expected_performance = False
+        self.visualize_expected_performance = True
 
         self.placeholder_char = '.'
         self.early_char = 'e'