@mkannwischer 2024-07-11
Maybe the above example wasn't a great example, but with the new visualization it's very easy to see when code is well scheduled. E.g., see the following dual-issuing code: 
```
                                                    // ----------------
                                                    // 0               
                                                    // |---------------
        sub v23.2S, v28.2S, v11.2S                  // *...............
        sub v21.2S, v28.2S, v1.2S                   // *...............
        sub v20.2S, v29.2S, v5.2S                   // .*..............
        sub v24.2S, v29.2S, v7.2S                   // .*..............
        add v21.2S, v0.2S, v21.2S                   // ..*.............
        add v27.2S, v10.2S, v23.2S                  // ..*.............
        add v24.2S, v6.2S, v24.2S                   // ...*............
        add v26.2S, v4.2S, v20.2S                   // ...*............
        add v4.2S, v4.2S, v5.2S                     // ....*...........
        add v5.2S, v6.2S, v7.2S                     // ....*...........
        sub v7.2S, v29.2S, v17.2S                   // .....*..........
        sub v25.2S, v29.2S, v13.2S                  // .....*..........
        add v6.2S, v12.2S, v13.2S                   // ......*.........
        sub v20.2S, v29.2S, v15.2S                  // ......*.........
 ```

It's also easier to compare to our [profiling scripts](https://github.com/slothy-optimizer/pqax/tree/main/tests/profiling) and LLVM MCA pipeline visualizations.