@hanno-becker 2024-07-11
I agree, this is easier to grasp than `// gap` annotations, which for models with large issue-rate are often dominating the output.